requirements.txt
wireless
twilio
python-dotenv
goprocam